move model definition to model/
